+++ tempest_fastapi_sdk/__init__.py
@@ -171,7 +171,7 @@ from tempest_fastapi_sdk.webpush import (
     WebPushSubscriptionSchema,
 )
 
-__version__: str = "0.24.0"
+__version__: str = "0.25.0"
 
 __all__: list[str] = [
     "CEP",